# utils/Swagger/oas.yaml
openapi: 3.0.3
servers:
  - url: 'https://test.es.sign.fiskaly.com/api/v1'
    description: |
      Entorno de prueba (Test)
  - url: 'https://live.es.sign.fiskaly.com/api/v1'
    description: |
      Entorno de producción (Live)
info:
  version: '1.13.0'
  title: fiskaly SIGN ES
  contact:
    email: 'info@fiskaly.com'
    name: 'Contacto'
    url: 'https://www.fiskaly.com/es'
  description: |
    Información Legal: [fiskaly.com/impressum](https://www.fiskaly.com/es/aviso-legal)
    <br>
    Política de Privacidad: [fiskaly.com/datenschutz](https://www.fiskaly.com/es/privacy-policy)
    <br>
    Solución: [fiskaly.com/signes](https://www.fiskaly.com/es/signes)
    <br>
    Guía para Desarrolladores: [developer.fiskaly.com/es/sign-es](https://developer.fiskaly.com/es/sign-es/introduccion)

    # Introducción

    **SIGN ES**, la [solución](https://www.fiskaly.com/es/signes) de fiskaly, ofrece una [RESTful](https://es.wikipedia.org/wiki/Transferencia_de_Estado_Representacional) API que asegura el cumplimiento con las normativas fiscales españolas como **TicketBAI** y **Verifactu**.

    **SIGN ES** proporciona un entorno de prueba para fines de integración y pruebas, así como un entorno de producción dedicado para realizar el envío de transacciones oficiales (facturas). Para acceder al entorno de producción, [contacta con nuestro equipo de ventas](mailto:sales@fiskaly.com).

...

    - tiene URLs orientadas a recursos,
...
    - usa verbos y códigos de estado HTTP estándar, y  
...

...

...

...

...

...

...
paths:
...
  '/clients/{client_id}':
    put:
      summary: |
        Crear cliente
      description: |
        Este endpoint se utiliza para crear un dispositivo cliente.

        Un dispositivo cliente permite identificar inequívocamente (a través del parámetro de ruta `device_id`) una terminal, sistema TPV, aplicación o cliente que emite facturas.

        Para emitir facturas necesitas un cliente por cada sistema TPV y al menos un dispositivo firmante.

        Esta llamada es idempotente.
      tags:
        - 'Clientes'
      security:
        - JWT: []
      operationId: createClient
      parameters:
        - $ref: '#/components/parameters/ClientIdentifier'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  $ref: '#/components/schemas/Client'
                metadata:
                  $ref: '#/components/schemas/Metadata'
              additionalProperties: false
      responses:
        '200':
          $ref: '#/components/responses/ClientResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedAccessResponse'
        '409':
          $ref: '#/components/responses/ResourceConflictResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
    patch:
      summary: Actualizar cliente
      description: 'Para actualizar el estado de un dispositivo cliente específico, este endpoint permite desactivar un dispositivo cliente de forma permanente y/o actualizar los metadatos del dispositivo cliente.'
      tags:
        - 'Clientes'
      security:
        - JWT: []
      operationId: updateClient
      parameters:
        - $ref: '#/components/parameters/ClientIdentifier'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: object
                  properties:
                    state:
                      $ref: '#/components/schemas/DeviceState'
                  additionalProperties: false
                metadata:
                  $ref: '#/components/schemas/Metadata'
              additionalProperties: false
      responses:
        '200':
          $ref: '#/components/responses/ClientResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedAccessResponse'
        '404':
          $ref: '#/components/responses/ResourceNotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
    get:
      summary: Obtener un cliente
      description: Obtiene un dispositivo cliente específico de la organización.
      tags:
        - 'Clientes'
      security:
        - JWT: []
      operationId: retrieveClient
      parameters:
        - $ref: '#/components/parameters/ClientIdentifier'
      responses:
        '200':
          $ref: '#/components/responses/ClientResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedAccessResponse'
        '404':
          $ref: '#/components/responses/ResourceNotFoundResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
  /clients:
    get:
      summary: Listar clientes
      description: Lista los dispositivos clientes de una organización.
      tags:
        - 'Clientes'
      security:
        - JWT: []
      operationId: listClients
      parameters:
        - $ref: '#/components/parameters/PaginationLimitQuery'
        - $ref: '#/components/parameters/PaginationTokenQuery'
      responses:
        '200':
          $ref: '#/components/responses/ClientsResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedAccessResponse'
        '500':
          $ref: '#/components/responses/InternalServerErrorResponse'
...
components:
  parameters:
...
    ClientIdentifier:
      description: 'Identifica un cliente de forma única con un identificador aleatorio en formato [UUIDv4](#section/Producto/Recursos).'
      in: path
      name: client_id
      required: true
      schema:
        $ref: '#/components/schemas/UniversallyUniqueIdentifierV4'
...
    PaginationLimitQuery:
      in: query
      name: limit
      required: false
      schema:
        $ref: '#/components/schemas/PaginationLimit'
    PaginationTokenQuery:
      in: query
      name: token
      required: false
      schema:
        $ref: '#/components/schemas/PaginationToken'
  responses:
...
    ClientsResponse:
      description: Devuelve una lista de clientes.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ClientResources'
    ClientResponse:
      description: Devuelve un cliente.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ClientResource'
...
    BadRequestResponse:
      description: Solicitud incorrecta (no válida)
      content:
        application/json:
          schema:
            type: object
            required:
              - content
            properties:
              content:
                $ref: '#/components/schemas/ErrorResponse'
            additionalProperties: false
    UnauthorizedAccessResponse:
      description: Acceso no autorizado.
      content:
        application/json:
          schema:
            type: object
            required:
              - content
            properties:
              content:
                $ref: '#/components/schemas/ErrorResponse'
            additionalProperties: false
    ResourceNotFoundResponse:
      description: El recurso solicitado no existe.
      content:
        application/json:
          schema:
            type: object
            required:
              - content
            properties:
              content:
                $ref: '#/components/schemas/ErrorResponse'
            additionalProperties: false
    ResourceConflictResponse:
      description: El recurso solicitado tiene un conflicto.
      content:
        application/json:
          schema:
            type: object
            required:
              - content
            properties:
              content:
                $ref: '#/components/schemas/ErrorResponse'
            additionalProperties: false
...
    InternalServerErrorResponse:
      description: Ocurrió un error interno del servidor
      content:
        application/json:
          schema:
            type: object
            required:
              - content
            properties:
              content:
                $ref: '#/components/schemas/ErrorResponse'
            additionalProperties: false
...
  schemas:
    ErrorResponse:
      description: Representación de respuestas de error.
      type: object
      required:
        - code
        - message
        - status
        - error
      properties:
        status:
          $ref: '#/components/schemas/ErrorStatus'
        code:
          $ref: '#/components/schemas/ErrorCode'
        error:
          $ref: '#/components/schemas/ErrorText'
        message:
          $ref: '#/components/schemas/ErrorMessage'
      additionalProperties: false
    ErrorStatus:
      description: Código de estado de error HTTP basado en números enteros.
      type: integer
      format: int32
      example: 401
    ErrorCode:
      description: Representación de un nombre de error único del código de estado de error HTTP.
      type: string
      example: E_UNAUTHORIZED_ACCESS
    ErrorText:
      description: Representación de una breve descripción del código de estado de error HTTP.
      type: string
      example: Unauthorized Access
    ErrorMessage:
      description: Representación de una descripción larga del código de estado de error HTTP.
      type: string
      example: Not authorized to perform this operation.
...
    Client:
      description: 'Representa un dispositivo de facturación al que accede una aplicación de facturación, ya sea que acceda o no a un servidor remoto para este propósito. Por ejemplo, en el caso de una aplicación basada en web, el número de serie del dispositivo de facturación será el del dispositivo en el que está instalado el navegador, y no el del servidor remoto al que accede para operar la aplicación.'
      type: object
      properties:
        signer_id:
          description: Identificador de un dispositivo firmante ya creado.
          oneOf:
            - $ref: '#/components/schemas/UniversallyUniqueIdentifierV4'
      additionalProperties: false
    ClientResources:
      description: Representación de los recursos de clientes.
      type: object
      required:
        - results
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/ClientResource'
        pagination:
          $ref: '#/components/schemas/Pagination'
      additionalProperties: false
    ClientResource:
      description: Representación del recurso cliente.
      type: object
      required:
        - content
      properties:
        content:
          $ref: '#/components/schemas/ClientDevice'
        metadata:
          $ref: '#/components/schemas/Metadata'
      additionalProperties: false
    ClientDevice:
      description: 'Representación de una abstracción de cliente TicketBAI y Verifactu proporcionando la información específica del cliente, así como la información de estado del cliente.'
      type: object
      required:
        - id
        - state
        - signer
      properties:
        id:
          $ref: '#/components/schemas/UniversallyUniqueIdentifierV4'
        state:
          $ref: '#/components/schemas/ClientState'
        signer:
          $ref: '#/components/schemas/ClientSigner'
      additionalProperties: false
...
    DeviceState:
      description: 'Cada dispositivo creado en el sistema se establecerá en el estado `ENABLED` (HABILITADO) y solo se puede establecer en `DISABLED` (DESHABILITADO), pero no al revés.'
      type: string
      enum:
        - ENABLED
        - DISABLED
    ClientState:
      description: 'Cada cliente en el sistema se establecerá en el estado `ENABLED` (HABILITADO) y solo se puede establecer como `DISABLED` (DESHABILITADO), pero no al revés.'
      type: string
      enum:
        - ENABLED
        - DISABLED
    ClientSigner:
      description: Representación de la información del dispositivo de firma para un cliente.
      type: object
      required:
        - id
      properties:
        id:
          $ref: '#/components/schemas/UniversallyUniqueIdentifierV4'
      additionalProperties: false
...
    UniversallyUniqueIdentifierV4:
      description: |-
        Un identificador único universal versión 4 (UUIDv4), es una etiqueta empleada para identificar objetos de forma única en
        sistemas informáticos. La representación de UUIDs está definida dentro del [RFC 4122](https://www.rfc-editor.org/rfc/rfc4122).
      type: string
      format: uuid
      pattern: '[a-f0-9]{8}-?[a-f0-9]{4}-?4[a-f0-9]{3}-?[89ab][a-f0-9]{3}-?[a-f0-9]{12}'
      example: 1c81cb86-c2e8-4074-afc3-a0601b2bf063
      x-go-type: uuid.UUID
      x-go-type-import:
        name: uuid
        path: github.com/google/uuid
...
    Metadata:
      description: |-
        Puede emplear este parámetro para adjuntar datos en formato clave/valor al objeto.  
        Los metadatos son útiles para almacenar datos estructurados que complementan el objeto.

        Se pueden adjuntar metadatos en los métodos de actualización de cada recurso. Si se incluye una clave con un valor vacío, se eliminará esa pareja clave/valor.

        Si se adjuntan metadatos a un objeto, estos serán devueltos en las respuestas a peticiones de ese objeto.

        **Nota**: Se pueden especificar hasta 20 claves, cada clave puede contener hasta 40 caracteres, y los valores pueden contener hasta 500 caracteres.
      type: object
      maxProperties: 20
      additionalProperties:
        type: string
        maxLength: 500
      example:
        my_property_1: '1234'
        my_property_2: 'https://my-internal-system/path/to/resource/1234'
...
    Pagination:
      description: 'Representa los datos de paginación de un endpoint de listado. Si no se especifican, los resultados actuales son la última página.'
      type: object
      required:
        - next
        - token
        - limit
      properties:
        next:
          $ref: '#/components/schemas/PaginationNext'
        token:
          $ref: '#/components/schemas/PaginationToken'
        limit:
          $ref: '#/components/schemas/PaginationLimit'
    PaginationNext:
      description: 'Representa la URL de petition la siguiente página de resultados. La petición está compuesta por el recurso de listado, el token de paginación, y el límite de resultados a presentar.'
      allOf:
        - $ref: '#/components/schemas/UniformResourceIdentifier'
    PaginationLimit:
      description: Representa el límite de resultados por página para un endpoint de listado.
      type: integer
      minimum: 1
      maximum: 100
      default: 10
    PaginationToken:
      description: 'Representa el token de paginación, que se emplea como puntero hacia la siguiente página de resultados.'
      type: string
      pattern: '^(?:[A-Za-z0-9+/]{4})*(?:[A-Za-z0-9+/]{4}|[A-Za-z0-9+/]{3}=|[A-Za-z0-9+/]{2}={2})$'
      maxLength: 1024
    UniformResourceIdentifier:
      description: Representa un identificador de recursos uniforme.
      type: string
...
  securitySchemes:
    JWT:
      description: |-
        JSON Web Token (JWT) empleado para control de acceso y autorización.
        - Ejemplo de uso: `Bearer <JWT>`
        - Ejemplo de encabezado HTTP: `Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`
      type: http
      scheme: bearer
      bearerFormat: JWT